+++ src/main.js
@@ -16,6 +16,7 @@ import aisLiveVesselsLayer from './data/aisLiveVessels.js';
 import militaryInstallationsLayer from './data/militaryInstallations.js';
 import militaryAwarenessLayer from './data/militaryAwareness.js';
 import moroccoPlacesLayer from './data/moroccoPlaces.js';
+import oncfTrainsLayer from './data/oncfTrains.js';
 import localDataLayers from './data/localLayers.js';
 import { LAYER_STATE_REGISTRY } from './data/layerState.js';
 import { registerDataCredits } from './data/dataCredits.js';
@@ -233,6 +234,7 @@ async function init() {
     dataManager.register(militaryAwarenessLayer);
     militaryAwarenessLayer.attachDataManager(dataManager);
     dataManager.register(moroccoPlacesLayer);
+    dataManager.register(oncfTrainsLayer);
     for (const layer of localDataLayers) {
       dataManager.register(layer);
     }